fix: shuttles list home page
